* commit 'a6e7f6068f8886b9e8a66a00b5efbc163c965ba4': (63 commits)
  more typos
  portage -> Portage at beginning of sentences
  missing period
  yet another typo
  typo
  fix formatting
  gary's tweak
  rao's wordsmithing
  mention version numbers and remove duplicate license/release.
  typo and comment on mispelling NCLUDE
  remove reference to MPI bindings
  ok im done playing with css because it hates me
  yet more tweaks
  more position tweaking
  account for filename changes
  trying more tweaks...
  finish text on examples
  more about AuxMeshTopology
  more tweaking
  start on example text
  ...

+++ jenkins/parseREADME.py
@@ -4,23 +4,28 @@ Please see the license file at the root of this repository, or at:
     https://github.com/laristra/portage/blob/master/LICENSE
 '''
 import subprocess, sys
-#This script parses the README file which describes how to build portage
-#It looks for code blocks enclosed in ``` and requires that the first
-#line of a block be a followed by a hostname.  The script will then
-#execute all commands specified after that point on the host machine
-#until the block is closed with a ```.
-#Args: Arg 1 is file to parse, Arg 2 is the workspace
+# This script parses the README file which describes how to build portage
+# It looks for code blocks enclosed in ``` and requires that the first
+# line of a block be a followed by a hostname.  The script will then
+# execute all commands specified after that point on the host machine
+# until the block is closed with a ```.
+# Args: Arg 1 is file to parse, Arg 2 is the workspace
 code = False
 scripts = {}
+mach_kw = "machine="
 for line in open(sys.argv[1]):
     if line.startswith("```"):
         code = not code
         currentScript = None
     elif code:
         if currentScript is None:
-            assert line.startswith("#")
+            # we can have code, and the code can start with comments,
+            # but we identify parts that should run on a particular machine
+            # with the "machine=" keyword
+            if not line.startswith("#") and mach_kw not in line: continue
+            host = line.split(mach_kw)[-1].strip()
             currentScript = ["set -e\n pwd \n cd "+ sys.argv[2] + "\n rm -Rf build \n"]
-            scripts[line.lstrip("#").strip()] = currentScript
+            scripts[host] = currentScript
         else: currentScript.append(line)
 
 retcode = 0
